[R] Formatting tweaks

+++ os/install_rstudio_server.sh
@@ -4,23 +4,43 @@
 # Set-up the rstudio user
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
-# Cook up a very long password
-declare -r rs_password="$(openssl rand -base64 50)"
-declare -r rs_user="brstudio"
+# If the user is already registered on the system, don't overwrite the existing
+# one!
 
-# Write here to copy into a key manager
-rm ~/rstudio_password
-echo $rs_password  >> ~/rstudio_password
+# This is really confusing and broken -- fix!
 
-# Create the user
-sudo useradd -mN $rs_user
 
-# With the password
-echo $rs_user:$rs_password | sudo chpasswd
 
-# Create the rstudio user group, and add them to it
-sudo groupadd rstudio
-sudo usermod -G rstudio $rs_user
+declare -r rs_usr=$(grep -q brstudio /etc/passwd)
+declare -r rs_grp=$(grep -q rstudio  /etc/group)
+
+if [ -z "$rs_usr" ] && [ -z "$rs_grp" ]
+then
+    # Cook up a very long password
+    declare -r rs_password="$(openssl rand -base64 50)"
+    declare -r rs_user="brstudio"
+
+    # Write here to copy into a key manager
+    rm ~/rstudio_password
+    echo $rs_password  >> ~/rstudio_password
+
+    # Create the user
+    sudo useradd -mN $rs_user
+
+    # With the password
+    echo $rs_user:$rs_password | sudo chpasswd
+
+    # Create the rstudio user group, and add them to it
+    sudo groupadd rstudio
+    sudo usermod -G rstudio $rs_user
+
+
+# ($left_bracket || $right_bracket) && ! ($left_bracket && $right_bracket)
+
+# If the commands are different (e.g. brstudio exists)
+elif ([ -z "$rs_usr" ] || [ -z "$rs_grp" ]) && ! \
+     ([ -z "$rs_usr" ] && [ -z "$rs_grp" ])
+
 
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 # Download and install the .deb file
@@ -29,15 +49,13 @@ sudo usermod -G rstudio $rs_user
 sudo apt-get install -y gdebi-core
 sudo apt-get install -y libapparmor1 # Required only for Ubuntu, not Debian
 
-# download .deb files
-
 # Get the latest version number, for 64 bit systems.
-curl http://www.rstudio.com/products/rstudio/download-server/ | \
-    grep -o 'https://download2.rstudio.org/rstudio-server.*amd64\.deb' >> \
-    rs_deb_url
+curl https://www.rstudio.com/products/rstudio/download-server/ |
+    grep -o 'https://download2.rstudio.org/rstudio-server.*amd64\.deb' >
+    rss_deb_url
 
-# Download the deb file
-paste rs_deb_url |  xargs wget -q
+# Downlo--quiet --directory-prefix ~deb file
+paste rss_deb_url |  xargs wget --quiet --directory-prefix ~
 
 # Run it. You really need to find a way for this to be accepted automatically (it prompts you to hit 'y')
 sudo gdebi rstudio-*-amd64.deb
